bench(c): C 参考实现与 .an 对齐规模与内存行为
C 现在是三态基线的基准, 因此两侧必须同算法、同规模、同内存行为:

- cd.c: 逐帧释放体素树/seen 树与 MotionList、被替换的 Vector3D、以及 state
  中移除的项, benchmark 结束时销毁 state。此前 `./cd 100 80` 峰值 RSS 约 1594 MB
  (YIAN 侧约 2-4 MB), 内存行为不可比; 修后约 2.4 MB, valgrind 0 泄漏 0 错误,
  输出仍为 344400 / 无参 4305。
- storage.c: Random 状态由 32 位改为 64 位, 与 .an 的 u64 LCG 抽样一致 (叶子元素
  总量此前差约 0.003%), 计数断言 21523360 不变。
- nbody.an: 热循环 distance 改用 std.num.f64.sqrt, 与 C 的 libm sqrt 同一原语
  (两侧 12M 步断言结果同为 -1)。

+++ bench/c/storage.c
@@ -13,20 +13,20 @@ typedef struct array_tree {
 } array_tree;
 
 typedef struct Random {
-    unsigned int stage;
+    unsigned long long stage;
 } Random;
 
 // 全局变量
 int count_arrays = 0;
 
 // 初始化随机数生成器
-void Random_init(Random* r, unsigned int seed) {
+void Random_init(Random* r, unsigned long long seed) {
     r->stage = seed;
 }
 
-// 生成下一个随机数
-unsigned int Random_next(Random* r) {
-    r->stage = (unsigned int)(r->stage * LCG_A + LCG_C);
+// 生成下一个随机数 (与 .an 基准同为 64 位状态, 保证叶子容量抽样一致)
+unsigned long long Random_next(Random* r) {
+    r->stage = r->stage * LCG_A + LCG_C;
     return r->stage;
 }
